backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
phonenumbers==8.13.47
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.4.0
redis==5.1.1
neo4j==5.24.0
httpx==0.27.2
jinja2==3.1.4
reportlab==4.2.2
tenacity==9.0.0
cachetools==5.5.0
python-multipart==0.0.10
pytest==8.3.3
